IR: migrate IrType utils about array element type.
